Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,8 @@ int main(){
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    float avg = float(a+b+c)/3
+    int avg = float(a+b+c)/3
+    
     printf("%f",avg)
     return 0;
 }
